Execute Playground code on a separate origin (#285)
Why

We want to execute Playground code on a separate origin, so that scripts are completely isolated from the main origin. In particular, scripts loaded through the "share" feature are not controlled by us, and often are written by another user to the person running it.

On a site that has access to e.g. sensitive user data or APIs on the main origin, this would be very important for security. That's why Playground warns in the console when it detects same-origin execution.

Our site is 100% static, so there's pretty limited risks -- an attacker could e.g. modify some page content, navigate the main window, etc. If we did something like integrate with GitHub gists for shorter Playground URLs, then an attacker might then be able to make authenticated GitHub requests through shared Playground code. So it seems like a good idea to close this hole before launch.

How

We have a new Cloud Run service called lit-dev-playground. Its URL is https://lit-dev-playground-5ftespv5na-uc.a.run.app/, and PR revisions have the same pr<PR>-<SHA> prefix that the main service has. When we go live, the URL will be https://playground.lit.dev

When we do an Eleventy build, the PLAYGROUND_SANDBOX environment variable is rendered into the HTML everywhere we instantiate a <playground-ide> or <playground-project> using the sandbox-base-url attribute. This attribute tells Playground to interact with a Service Worker at that URL, instead of the default one on the same origin.

The server has a new MODE environment variable switch. When it's set to playground, it only serves the js/ directory, instead of the full site. It