Fix check external link test timing out (#438)
The broken link checker has two behaviors when parsing a page. For each link we check whether or not it is broken. Then we also decide whether or not to queue up that link to check its links. This second behavior is changed by the --exclude-internal flag. Previously the localhost:8080/slack-invite/ link was marked as internal so we don't check to see if it is broken, but we do queue it so that we can check if it contains external links that are broken. This results in treating the the slack domain as internal and checking for broken external links. Now the crawler uses the current domain (slack) to decide what is an internal link. This leads to the crawler walking slack. This eventually leads to another redirect which continues for hours.

Note in the example above I use /slack-invite/ url, which we have set up to redirect to slack.com. This tricks the crawler to treat slack as an internal domain.

Co-authored-by: Andrew Jakubowicz <ajakubowicz@google.com>