Add new lit-dev-tools-esm package
The lit-dev-tools package is currently CommonJS, because mostly it
is used for Eleventy plugins, and Eleventy doesn't support ES
modules (https://github.com/11ty/eleventy/issues/836).

We want ES modules for this new redirect checker script, because it
needs to import some ES modules, and that is difficult to do with
TypeScript, because TypeScript doesn't allow emitting an actual
`import` statement, which is how CommonJS -> ESM interop works
(https://github.com/microsoft/TypeScript/issues/43329).

We also an't really have a mix of CommonJS and ESM in the same package,
because the {"type": "module"} field has to be set. We could use
.mjs extensions, but TypeScript won't emit those.

So the simplest solution seems to be to just have two packages!