Preserve viewport-relative position of samples when language preference is changed (#483)
Code samples commonly have a different number of lines in TypeScript vs JavaScript. This can cause page layout shifts when the language changes, especially because changing the language for one sample changes it for all other samples on the page too. These shifts are very visually distracting, and can even cause the active code sample to move entirely out of view.

This change mitigates that by comparing the viewport-relative position of the particular instance of the language switch that was interacted with before and after the change, and re-scrolling the window so that the switch remains in the same location visually.