Fix broken folding range
In this example, the /* playground-fold */ comment was associated
with the decorators import statement. Since that import gets
removed as part of our transform, the comment was getting removed
too.

A quick fix here is to just shift the comment down. This means the
decorator import is now visible in TS mode, but that seems fine.