Run a separate dev server for the playground sandbox (#519)
Previously, during local development, we ran the Playground sandbox on the same origin as the main site for simplicity. On production, though, we run a separate playground.lit.dev server for the sandbox. We do this in production for security, because we want to run untrusted code in an isolated context.

Now during local dev we launch two @web/dev-servers: one for the main site, and one for the playground sandbox, mirroring what we do on production.

The ports are also now static: 5415 for the main dev server, and 5416 for the playground dev server. Previously the dev server opened on the first port that was available starting from 8000. We need static ports because Eleventy now needs to know what the Playground sandbox URL will be, since it is baked into the HTML it generates. I picked these numbers mostly arbitrarily.

I've also done the equivalent for the "local production" configuration (npm start), and updated the docs for Docker. So there's now a port scheme that is hopefully easy to remember:

Local dev server: port 5415 / 5416
Local production server: port 6415 / 6516
Local production server using docker: port 7415 / 7416
The main motivation for this was upcoming work to set up a CSP policy (#517). We need a different CSP policy on the main vs playground sandbox servers, because the main server must not run untrusted code, while the playground sandbox must be able to.