Implement a Content Security Policy (CSP) for lit.dev (report only for now) (#526)
Adds a new contentSecurityPolicyMiddleware which we run in both the prod and dev servers.

For now it is using the Content-Security-Policy-Report-Only heading. This means violations will be shown in the console, but they won't actually throw any exceptions. On the production site, violations will additionally get reported to the Google CSP collector, which has a dashboard we can use to see if violations are occuring.

The policy isn't as strict as we would like. In particular, we have to allow unsafe-eval because of a library we are using inside the Playground web worker. More details in the TODO. I plan to send some followup PRs that will incrementally lock down this policy.

Part of #517