Replace onclick handler with inlined script (whose hash is automatically added to CSP) (#554)
Turns out there was a CSP error that I missed that only affected mobile view. An inline onclick handler on the mobile nav open button.

This is a quick hack to roll forward that turns the onclick handler into an inlined script, which gets automatically hashed for CSP. Better solution would probably be to make a custom element that contains both the button and the drawer, and hook things up with a lit template.