Restore target:esnext to samples tsconfig (#582)
Fixes a regression from #560 where we accidentally dropped from "target": "esnext" to "target": "es2020" in the tsconfig.json for sample files (I made a common tsconfig.base.json which defaults to es2020 but forgot to override that in the samples one). We want esnext output for samples because we want to emit standard static class fields (note that @lit/ts-transformers switches instance class field initializers to be constructor-initialized if they are decorated with @property, to avoid that incompatibility).