Show banner and version selector for old docs (#599)
Adds a "You're viewing docs for an older version of Lit" banner to the top of the `/v1/` docs, and a Version selector in the top-left.

Version selector is hidden behind the `?mods=old-docs` mod for now.

Also fixes a CSS bug where the margin above the LHS nav was wrong when the content was too short (affected the first page of the old docs, and also the Community page of the new docs).

Currently when we switch versions, we always jump to the first page of the docs. In a follow up PR, I'll make it jump to the more specific matching page.