Make closing the gist flyout more reliable (#612)
You can now close the flyout by clicking the "Share" button again. Previously this didn't work because clicking on it 1. caused the flyout to close itself (because it closes itself when it detects a click outside of itself), and then 2. the share button to open the flyout again (because it naively opened for every click).

You can now close the flyout by clicking on the preview. Previously this didn't work because those clicks went to the iframe window, and never reached the host window. Now we temporarily set pointer-events: none when the flyout is open so that the main window gets those events instead.