Make sample generation more robust
Previously we used chokidar (a filesytem watcher) to find all of the
files that needed to be symlinked from samples/* into samples/js -- both
in the normal one-time mode, and for watch mode.

In normal mode, we waited for chokidar's "ready" event, and then assumed
that all files would have been detected. Although this does seem to be
what the documentation suggests the "ready" event means, we were seeing
cases where only a subset of files were getting detected. I suspect
this is because on some machines, the "ready" event is not a reliable
indicator that all files have been found.

We now instead do a more straightforward glob match for files when we're
in the normal one-time mode, which should avoid any weirdness around
chokidar. We still use chokidar when in --watch mode.