Refactor playground page into a component (#710)
Previously, most of the logic for orchestrating the Playground page was in a regular script that waited for DOMContentLoaded, and then manually queried for the underlying Playground components and hooked them together.

Now... we still do most of that! Except, it's contained within a new <litdev-playground-page> component.

This is the first step in a refactoring to make the Playground easier to work on.