Update getting-started.md
Exactly *how* to directly import (without a build step) is not at all clear in this documentation, a serious hindrance to developers wanting as easy as possible a way to test the waters.  Proposing this explicit line to make things much clearer, since the remaining documentation here all assumes you'll do a build step.