Add version links so that v2 <-> v3 changing doesnt change page.
A user on v2 docs for a certain page should remain on that page when
switching to v3 and vice-versa. This change makes it so switching
between v2 and v3 will keep you on the same page, but provide the
different versions of the page.

Tested manually on every single page.