Release Lit v3 pre-release docs with blog post (#1116)
* Copy over all unversioned content into a versioned directory

The only file changed in the move is the root `docs.json` file. Instead
a `v2.json` file was added.

* Add initial build-unversioned-docs script.

This copies over and adds the correct permalinks to the selected
version content. Currently moving `v2` -> `unversioned`.

* Fix wireit so npm run dev doesnt spin infinitely.

* Add rel=canonical link from latestVersion to unversioned page.

* Make all authored cross links versioned.

Unversioned links will be generated automatically by the generated
unversioned pages.

* Add fixUnversionedCrossLinks so unversioned cross-linking work.

This works by piping all the content through a function that will
fix cross links on unversioned pages.

* Add integration tests checking added features.

* unblock failing link by adding it to the known good list.

* Fix api shortcode such that it uses a versioned URL

The versioned URL is stripped for unversioned URLs by our tooling.

Add a test to confirm this behavior.

* Apply code review feedback. Thank you!

* Code review feedback from Justin

 - Replaced sync fs with async fs.
 - Made lightweight YAML parser more robust using regex.

* Add undiscoverable v3 generated documentation.

Currently the Lit 3 commit being used is
d04a3e30eb3ae3520fb0ac163fb5ddbbf6030620 because
https://github.com/lit/lit/pull/3894 fixes building Lit 3.

API docs are visible at /docs/api/v3/

* Explicitly bump node heap size.

This will hopefully avoid the Github Actions failures.

* Add types/codemirror to fix TS 5 error.

Thank you Augustine! Re-ran generator with no content changes.

* Remove unneeded dependency

Co-authored-by: Augustine Kim <augustinekim@google.com>

* Code review feedback with Wireit env.

* Fix formatting & filter out v3/api/index.html page

* Copy over v2 docs to v3 unchanged.

* Add banner and config

* Fix all cross links in v3 docs to be v3 specific.

 - Find replace all /v2/ with /v3/
 - Replace api shortcode with api-v3 shortcode - linking to v3 api docs

* Add Lit 3.0 upgrade guide - first attempt.

* Add v3 to the Lit.dev dropdown.

* Upstream ssr client-usage docs to v3 page.

This makes the page a duplicate of the v2 page. Git for some reason did
not include these changes when merging from main.

* Remove pre-release v3 docs from search index

* Add version links so that v2 <-> v3 changing doesnt change page.

A user on v2 docs for a certain page should remain on that page when
switching to v3 and vice-versa. This change makes it so switching
between v2 and v3 will keep you on the same page, but provide the
different versions of the page.

Tested manually on every single page.

* Update tools section for v3 (#1118)

* Update tools section for v3

* Better wording for versions tested

* Add pre-releases banner content to v3 banner.

* Update packages/lit-dev-content/site/docs/v3/releases/upgrade.md

Co-authored-by: Augustine Kim <augustinekim@google.com>

* Update packages/lit-dev-content/site/docs/v3/releases/upgrade.md

Co-authored-by: Augustine Kim <augustinekim@google.com>

* Lit 3.0 pre-releases blog post (#1115)

* Lit 3.0 pre-releases blog post

* Apply suggestions from code review

* Add link to upgrade guide

---------

Co-authored-by: Augustine Kim <augustinekim@google.com>

* Remove a small paragraph that didn't really make sense from upgrade.

* Add missing word "can".

* Fix blog dropdown referring 3.0 -> v3

---------

Co-authored-by: Augustine Kim <augustinekim@google.com>
Co-authored-by: Justin Fagnani <justinfagnani@google.com>