[search] update algolia version and send clicks
Updated algolia search version, and introduced search-insights for anonymized click analytics. Specifically does not use cookies and does not store any identifiable data. Reason for introduction is because I realized I am only updating our site based on searches with no results, but not searches with no clicks.

This causes the problem that we had results for `componentDidMount` which linked to this page https://lit.dev/docs/frameworks/react/#how-it-works which is a bad result. Something like that could have been caught with click data.